Bug squashing related to editor replacement
